Moved scalacheck integration code to org.scalatest.check package, and deprecated them in prop package.
